+++ src/SlmMail/Service/MandrillService.php
@@ -107,6 +107,21 @@ class MandrillService extends AbstractMailService
     }
 
     /**
+     * Get all the information about a message by its Mandrill id
+     *
+     * @link https://mandrillapp.com/api/docs/messages.JSON.html#method=info
+     * @param  string $id
+     * @return array
+     */
+    public function getMessageInfo($id)
+    {
+        $response = $this->prepareHttpClient('/messages/info.json', array('id' => $id))
+                         ->send();
+
+        return $this->parseResponse($response);
+    }
+
+    /**
      * ------------------------------------------------------------------------------------------
      * USERS
      * ------------------------------------------------------------------------------------------
@@ -699,6 +714,7 @@ class MandrillService extends AbstractMailService
      * @param  HttpResponse $response
      * @throws Exception\InvalidCredentialsException
      * @throws Exception\ValidationErrorException
+     * @throws Exception\UnknownTemplateException
      * @throws Exception\RuntimeException
      * @return array
      */
@@ -717,6 +733,10 @@ class MandrillService extends AbstractMailService
                 ));
             case 'ValidationError':
                 throw new Exception\ValidationErrorException(sprintf(
+                    'An error occurred on Mandrill (code %s): %s', $result['code'], $result['message']
+                ));
+            case 'Unknown_Template':
+                throw new Exception\UnknownTemplateException(sprintf(
                     'An error occurred on Mandrill (code %s): %s', $result['code'], $result['message']
                 ));
             default: