+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 1.0.1
+
+- Add the new "getMessageInfo" to Mandrilll service.
+- Add a new "UnknownTemplateException" for services that support templates and report this error (currently, only Mandrill)
+
+## 1.0.0
+
+- Doc update
+
 ## 1.0.0 RC3
 
 - Add exceptions support for Amazon SES.